Code to read a csv file and convert it to an html table
What: This is code to read a csv file and generate an html table of the data
Why: My ambition is to make a product flier for East Wind products.  The flier contains a large table with all of the products East Wind makes.  The products occasionally change, so I want to make it easy to change the table by changing the data.  

Eventually, want an east winder to go to a webpage, adjust the data, and be able to print out a new flier.  I am going to construct the flier with html markup and a css print stylesheet.