hard-coded csv file and deleted unnecessary file-reading functions
What: I deleted everything that had to do with reading csv files and hard-coded an example csv file. 
Why:  I am going to create a text-box in which you are supposed to paste the csv file instead of reading one from a file system.  This will be easier for other people to understand, and the code will actually work, since the import code in the previous commit was broken and I didn't want to take the time to figure out how to fix it.