Added texteara for csv input
What:  Created textarea so that a person can copy and paste a csv file into it.  The csv will be used to auto-generate a table of the data. 
Why:  I am eventually going to use CSS stylesheets to create a print-ready pamphlet of bulk nut-butters products.  Coding a table by hand with all of the data is boring when I could have javascript auto-generate the table for me by reading a csv file.  Eventually, I am going to leave behind an html page where a person can change the data in the data in the form and hit the 'print' button, and get a nice-looking pamphlet.