Added a product summary to the pamphlet
What: I added a product summary to the pamphlet.  I created a new field in the form for the summary.  I added code to auto-generate html of the summary for the pamphlet.  The pamphlet html nows consists of two divs: the summary div and the table div. 
Why:  I want to be able to edit the summary without editing the html directly.  

Silly coding decisions: the "parser" for the summary just looks for a line-break between paragraphs. It is the dumbest parser possible.  In the future, I should swap that out for a legitimate markdown parser or something.