Updated css pertaining to table
What: I made the table stretch to fit nearly the whole box set aside for the table
Why: it looks good of the table takes up the whole lower left box in the flyer.  

Errors I ran into: apparently, the page size I am printing on is not actually 11 inches.  The box holding the table is 6 inches wide and 9 inches tall, but the box is running off the bottom edge, because when I stretch the table vertically to fill all nine inches, the table runs off the bottom of the page.  
Also, CSS is confusing.