Added SVG Logo To Pamphlet
What: I added an SVG logo to the header section of the pamphlet
Why:  Why did I do an SVG logo instead of an image?  And why is it just the words "East Wind Nut Butters" with no fancy styling?  The first reason is that I'm making this page a pamphlet-generating app, and I want it all in one file, so that if a person saves this page offline, everything you need to run this app will be in one file.  If i used external images, I would have a bunch of extra images hanging around.  
As for why the logo is so basic...Pull up a chair and let me tell you a story...
Because I was hand-coding it in Vim, that's why.  But Tara, you say, why are you hand-coding svg in vim?
1. I can't find the original logo files for our labels, so I can't grab something from that. 
2. Photoshop and illustrator don't work, because we recently upgraded to El Capitan and don't have the right Java runtime for them.  
3. I'm at the end of a very slow internet connection, so downloading the javascript runtime takes an hour.
4. Same reason I can't download inkscape.
5. Gimp apparently takes 1 hour to load fonts.
...so. I decided to have a race.  Could I create a logo with vim and hand-coding SVG faster than I could get any of my graphics programs to work?  The answer is yes...yes I could.