Fixed CSV 'parser' bug: splitting quoted field with commas into separate cells
What:  My so-called parsing function was splitting a quoted field that looked like this: "G, F, IOW, 6"  into four separate cells because it did not have any way of recognizing that the quotes made it one field.  My CSV parser does line-by-line parsing.  The old code was parsing the line 

88888, Smooth w/salt  , 35lb tub , 1  , 0-00000-000000-2, "G, TC, IOW"

incorrectly.  The new code uses a regex to split that line on every comma that is not in a set of quotes.

This regex will NOT correctly read CSV files with double quotes.  It is still a hack. It is NOT full-featured fix.  

Why:  Why did I go with this solution instead of doing it the 'right' way and finding a real javascript CSV parser?  Because I am 'spiking.'  Spiking is exploring a solution without really knowing if it will work or not. Also, I have lots of experience with debugging regexes.  I have very little experrience debugging full-blown parsers, so the less code I have to decipher, the better.  This solution, though not ideal, has the advantage of only requiring six lines of code for me to understand, instead of reading through someone's heavyweight library. Also, I am trying to keep this web app all on one page and fairly lightweight.  Adding a ton of library code would not help with that ambition.