Changed table generation rule to have single-cell columns span all 6 columns
What:  The table display was not working.  There were some table rows that had only one cell with a lot of text.  Since the css rules declared that the text should be all on one line, thee first column was really big and squeezed the rest of the columns off the printed page.  I took the table rows with only one cell and used the colspan attribute to stretch them across the entire table. This meant that the first column got a lot skinnier.
Why: It made the table look better on the printed page.